Update config.py
We were creating a connection multiple times. There is no need to do it in this file too.
Also, we needed a way to call the same files on our own computers but the password variable is different for each obviously. We solved this by simply taking the password as user input.