Update app.py
In the clubs table, I replaced domestic_competition_id with the competition_name which is much more understandable for users. For this, in the query, I joined clubs table and competitions table while only getting the competition names with "domestic_league" type to avoid duplicated clubs. (every club plays in multiple competitions)